refactor(langfuse): replace `LangfuseBootstrap` with `LangfuseProvisioner`
- Renamed `LangfuseBootstrap` to `LangfuseProvisioner` for improved clarity and extensibility.
- Updated Langfuse provisioning to include error-tolerant steps and enhanced model discovery.
- Replaced occurrences of `LangfuseBootstrap` in `lifetime_manager.py` and `AgentEndpointsDiscoveryService`.
- Updated related dependencies and method calls for consistency.
