+++ aihub_api/aihub_api/runners/lifetime/lifetime_manager.py
@@ -4,7 +4,7 @@ from contextlib import asynccontextmanager
 
 import boto3
 from aihub_lib.infrastructure.api.AIHubSettings import AIHubSettings
-from aihub_lib.infrastructure.langfuse.LangfuseBootstrap import LangfuseBootstrap
+from aihub_lib.infrastructure.langfuse.LangfuseProvisioner import LangfuseProvisioner
 from aihub_lib.infrastructure.milvus.MilvusSettings import MilvusSettings
 from aihub_lib.infrastructure.mongo.MongoSettings import MongoSettings
 from aihub_lib.infrastructure.nats.NatsSettings import NatsSettings
@@ -34,30 +34,16 @@ from aihub_api.sockets.sender.WebSocketSender import WebSocketSender
 logger = logging.getLogger(__name__)
 
 
-def _create_langfuse_bootstrap() -> LangfuseBootstrap | None:
+def _create_langfuse_provisioner() -> LangfuseProvisioner | None:
     """
-    Create a LangfuseBootstrap instance, returning None if Langfuse is not configured.
+    Create a LangfuseProvisioner instance, returning None if Langfuse is not configured.
     This prevents missing environment variables from crashing API startup.
     """
     try:
-        return LangfuseBootstrap()
+        return LangfuseProvisioner()
     except Exception as e:
         logger.warning(f"Langfuse not configured (non-fatal): {e}")
         return None
-
-
-async def _bootstrap_langfuse(langfuse_bootstrap: LangfuseBootstrap) -> None:
-    """
-    Bootstrap Langfuse with AI-Hub LLM connections and evaluators.
-
-    This runs after all core services are initialized so that Langfuse
-    can call back to AI-Hub's OpenAI-compatible endpoint during experiments.
-    Errors are logged but don't block API startup.
-    """
-    try:
-        await langfuse_bootstrap.bootstrap()
-    except Exception as e:
-        logger.warning(f"Langfuse bootstrap failed (non-fatal): {e}")
 
 
 @asynccontextmanager
@@ -199,7 +185,7 @@ async def lifetime_manager(app: FastAPI) -> AsyncGenerator:
 
         api_app = app.state.api_app
 
-        langfuse_bootstrap = _create_langfuse_bootstrap()
+        langfuse_provisioner = _create_langfuse_provisioner()
 
         if hasattr(api_app.state, "agent_controller"):
             agent_discovery_service = AgentEndpointsDiscoveryService(
@@ -207,7 +193,7 @@ async def lifetime_manager(app: FastAPI) -> AsyncGenerator:
                 api_app=api_app,
                 controller=api_app.state.agent_controller,
                 locale_handler=ApiLocaleHandler(),
-                langfuse_bootstrap=langfuse_bootstrap,
+                langfuse_provisioner=langfuse_provisioner,
                 discovery_interval=60,  # Check for new agents every 60 seconds
             )
             await agent_discovery_service.start()
@@ -231,9 +217,9 @@ async def lifetime_manager(app: FastAPI) -> AsyncGenerator:
         await initialize_roles()
         await initialize_knowledge_buckets()
 
-        # Bootstrap Langfuse with AI-Hub LLM connections
-        if langfuse_bootstrap:
-            await _bootstrap_langfuse(langfuse_bootstrap)
+        # Provision Langfuse with AI-Hub LLM connections
+        if langfuse_provisioner:
+            await langfuse_provisioner.provision()
 
         # Yield control back to FastAPI to start serving requests
         yield
